+++ km_api/mailing_list/__init__.py
@@ -1,4 +1,0 @@
-"""App for managing mailing lists.
-"""
-
-default_app_config = 'mailing_list.apps.MailingListAppConfig'