+++ docs/changelog.rst
@@ -6,11 +6,12 @@ Changelog
 In Development
 **************
 
-Breaking Changes
-  * Completely changed the way that settings are provided. Rather than being specified in a separate python file that is then imported, behavior is now set through environment variables.
-
 Features
   * :issue:`379`: Add endpoint to retrieve/update a user's Know Me subscription through Apple.
+
+Miscellaneous
+  * Completely changed the way that settings are provided. Rather than being specified in a separate python file that is then imported, behavior is now set through environment variables.
+  * :issue:`383`: Removed MailChimp integration.
 
 
 ******