+++ docs/changelog.rst
@@ -11,7 +11,8 @@ Features
 
 Miscellaneous
   * Completely changed the way that settings are provided. Rather than being specified in a separate python file that is then imported, behavior is now set through environment variables.
-  * :issue:`383`: Removed MailChimp integration.
+  * :issue:`383`: Remove MailChimp integration.
+  * :issue:`384`: Remove CloudWatch integration.
 
 
 ******